Implemented time limit using command line argument

+++ src/clargs.rs
@@ -37,5 +37,5 @@ pub struct Args {
 
     /// Maximum time (in seconds) that a single plugin can run before moving on to the next one. No time limit by default.
     #[arg(short, long)]
-    pub time_limit: Option<u32>,
+    pub time_limit: Option<u64>,
 }
